Don't try to avoid crossing service roads.
Service roads are impassable in the AAA methodology, and LTS 4 in our implementation of their methodology. This meant that roads that crossed service roads (e.g. the park parking lot entrance in the added test) were weighted heavily, as if they were crossings of major arterials. This adds a new LTS 5 category, which is equivalent to LTS 4 except that it doesn't affect intersections when neither of the streets actually being used are service roads.

Also, add automated test to confirm this fixed the parking lot issue initially observed